pyproject.toml
[build-system]
requires = ['flit_core>=3.9']
build-backend = 'flit_core.buildapi'

[project]
name = 'time-lapse'
version = '8.0'
description = 'Time-lapse movie assembly'
readme = 'README.md'
requires-python = '>=3.13'
license = {file = 'LICENSE'}
authors = [
    {name = 'Arne de Laat', email = 'arne@delaat.net'},
]
maintainers = [
    {name = 'Arne de Laat', email = 'arne@delaat.net'},
]
keywords = [
    'ffmpeg',
    'photography',
    'time-lapse',
]
classifiers = [
    'Intended Audience :: Developers',
    'Intended Audience :: Other Audience',
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Topic :: Multimedia :: Video :: Conversion',
]
dependencies = [
    'ffmpeg-python==0.2.0',
    'pillow==11.0.0',
]

[project.optional-dependencies]
graph = [
    'graphviz',
]
dev = [
    'coverage==7.6.9',
    'mypy==1.14.0',
]
publish = [
    'flit==3.10.1',
]

[project.urls]
Homepage = 'https://arne.delaat.net/timelapse.html'
Repository = 'https://github.com/153957/time-lapse'

[project.scripts]
timelapse = 'time_lapse.cli:timelapse'
detect_audio = 'time_lapse.cli:detect_audio'

[tool.mypy]
ignore_missing_imports = true
show_column_numbers = true
show_error_codes = true
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 120
target-version = 'py311'

[tool.ruff.format]
quote-style = 'single'  # Prefer single quotes, except for triple quotes strings

[tool.ruff.lint]
select = ['ALL']  # https://docs.astral.sh/ruff/rules/
ignore = [
    'CPY001',  # Do not require copyright notices
    'D104',  # Allow missing package docstring
    'D203',  # Should instead be no blank line before class docstring
    'D213',  # Summary should instead be on the first line
    'EM',  # Allow messages directly in exceptions
    'FBT001',  # Allow positional for boolean arguments
    'FBT002',  # Allow default value for boolean arguments
    'PD',  # Not using pandas
    'PLR0913',  # Allow functions with many arguments
    'PT',  # Not using pytest
    'RET504',  # Allow variable assignment before return
    'SIM108',  # Allow if-else block instead of ternary
    'T201',  # Allow using print
    'TRY003',  # Specific messages for common exception classes
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = 'single'

[tool.ruff.lint.isort]
lines-between-types = 1
section-order = [
    'future',
    'standard-library',
    'third-party',
    'ffmpeg',
    'first-party',
    'local-folder',
]

[tool.ruff.lint.isort.sections]
ffmpeg = ['ffmpeg']

[tool.coverage.run]
branch = true
source = [
    'time_lapse',
    'tests',
]
